fixed task interactor for update and updateTask added
