+++ app/graphql/types/inputs/update_task_input.rb
@@ -1,0 +1,9 @@
+module Types
+  module Inputs
+    class UpdateTaskInput < Types::BaseInputObject
+      argument :task_id, ID, required: true
+      argument :title, String, required: false
+      argument :description, String, required: false
+    end
+  end
+end
