+++ app/graphql/types/mutation_type.rb
@@ -1,5 +1,6 @@
 module Types
   class MutationType < Types::BaseObject
     field :create_project, mutation: Mutations::CreateProject
+    field :update_task, mutation: Mutations::UpdateTask
   end
 end
