+++ app/graphql/types/payloads/update_task_payload.rb
@@ -1,0 +1,8 @@
+module Types
+  module Payloads
+    class UpdateTaskPayload < Types::BaseObject
+      field :task, TaskType, null: true
+      field :errors, [Types::UserError], null: false
+    end
+  end
+end
